+++ src/Units/Emus.cs
@@ -4,7 +4,7 @@ namespace ShapeCrawler.Units;
 
 public readonly ref struct Emus(long emus)
 {
-    public decimal AsPoints() => Math.Round(emus / 12700m, 2);
+    public double AsPoints() => Math.Round(emus / 12700d, 2);
 
-    public decimal AsPixels() => Math.Round(emus / 9525m, 2);
+    public double AsPixels() => Math.Round(emus / 9525d, 2);
 }
